create mock for shared code for testing
