+++ test/actions/api/linodes.spec.js
@@ -1,26 +1,11 @@
 import sinon from 'sinon';
-import { expect } from 'chai';
-import * as fetch from '~/fetch';
 import {
   powerOnLinode,
   powerOffLinode,
-  rebootLinode
+  rebootLinode,
+  UPDATE_LINODE
 } from '../../../src/actions/api/linodes';
-import * as linode_actions from '../../../src/actions/api/linodes';
-
-const mock_context = async (f, rsp, state={}) => {
-  const auth = { token: 'token' };
-  let getState = sinon.stub().returns({
-    authentication: auth,
-    ...state
-  });
-  let dispatch = sinon.spy();
-  let fetchStub = sinon.stub(fetch, "fetch").returns({
-    json: () => rsp
-  });
-  await f({auth, getState, dispatch, fetchStub});
-  fetchStub.restore();
-};
+import { mock_context } from '../../mocks';
 
 describe("actions/linodes/power", sinon.test(() => {
   it('returns linode power boot status', async () => {
@@ -34,11 +19,11 @@ describe("actions/linodes/power", sinon.test(() => {
       sinon.assert.calledWith(fetchStub,
         auth.token, '/linodes/foo/boot', { method: "POST" });
       sinon.assert.calledWith(dispatch, {
-        type: linode_actions.UPDATE_LINODE,
+        type: UPDATE_LINODE,
         linode: { id: "foo", state: "booting" }
       });
     }, {
-      type: linode_actions.UPDATE_LINODE,
+      type: UPDATE_LINODE,
       linode: { id: "foo", state: "booting" }
     });
   });
@@ -54,11 +39,11 @@ describe("actions/linodes/power", sinon.test(() => {
       sinon.assert.calledWith(fetchStub,
         auth.token, '/linodes/foo/shutdown', { method: "POST" });
       sinon.assert.calledWith(dispatch, {
-        type: linode_actions.UPDATE_LINODE,
+        type: UPDATE_LINODE,
         linode: { id: "foo", state: "shutting_down" }
       });
     }, {
-      type: linode_actions.UPDATE_LINODE,
+      type: UPDATE_LINODE,
       linode: { id: "foo", state: "shutting_down" }
     });
   });
@@ -74,11 +59,11 @@ describe("actions/linodes/power", sinon.test(() => {
       sinon.assert.calledWith(fetchStub,
         auth.token, '/linodes/foo/reboot', { method: "POST" });
       sinon.assert.calledWith(dispatch, {
-        type: linode_actions.UPDATE_LINODE,
+        type: UPDATE_LINODE,
         linode: { id: "foo", state: "rebooting" }
       });
     }, {
-      type: linode_actions.UPDATE_LINODE,
+      type: UPDATE_LINODE,
       linode: { id: "foo", state: "rebooting" }
     });
   });
